// src/apis/openapi/openapi-types.ts
export namespace OpenAPI {
  // OpenAPI extensions can be declared using generics
  // e.g.:
  // OpenAPI.Document<{
  //   'x-amazon-apigateway-integration': AWSAPITGatewayDefinition
  // }>
  export type Document<T extends {} = {}> =
    | OpenAPIV3.Document<T>
    | OpenAPIV3_1.Document<T>
  export type Operation<T extends {} = {}> =
    | OpenAPIV3.OperationObject<T>
    | OpenAPIV3_1.OperationObject<T>
  export type Parameter =
    | OpenAPIV3_1.ParameterObject
    | OpenAPIV3.ParameterObject
  export type Parameters =
    | OpenAPIV3_1.ParameterObject[]
    | OpenAPIV3.ParameterObject[]
  export type PathsObject = OpenAPIV3.PathsObject | OpenAPIV3_1.PathsObject
  export type PathItemObject =
    | OpenAPIV3.PathItemObject
    | OpenAPIV3_1.PathItemObject
  export type ResponseObject =
    | OpenAPIV3.ResponseObject
    | OpenAPIV3_1.ResponseObject
  export type RequestBodyObject =
    | OpenAPIV3.RequestBodyObject
    | OpenAPIV3_1.RequestBodyObject
  export type SchemaObject = OpenAPIV3.SchemaObject | OpenAPIV3_1.SchemaObject
  export type ArraySchemaObject =
    | OpenAPIV3.ArraySchemaObject
    | OpenAPIV3_1.ArraySchemaObject
  export type SchemaObjectType =
    | OpenAPIV3.ArraySchemaObjectType
    | OpenAPIV3.NonArraySchemaObjectType
    | OpenAPIV3_1.ArraySchemaObjectType
    | OpenAPIV3_1.NonArraySchemaObjectType
  export type SecuritySchemeObject =
    | OpenAPIV3.SecuritySchemeObject
    | OpenAPIV3_1.SecuritySchemeObject

  export type HttpSecurityScheme =
    | OpenAPIV3.HttpSecurityScheme
    | OpenAPIV3_1.HttpSecurityScheme
  export type ApiKeySecurityScheme =
    | OpenAPIV3.ApiKeySecurityScheme
    | OpenAPIV3_1.ApiKeySecurityScheme
  export type OAuth2SecurityScheme =
    | OpenAPIV3.OAuth2SecurityScheme
    | OpenAPIV3_1.OAuth2SecurityScheme

  export type SecurityRequirementObject =
    | OpenAPIV3.SecurityRequirementObject
    | OpenAPIV3_1.SecurityRequirementObject

  export type ServerObject = OpenAPIV3.ServerObject | OpenAPIV3_1.ServerObject

  // Export enum httpmethods
  export enum HttpMethods {
    GET = 'get',
    PUT = 'put',
    POST = 'post',
    DELETE = 'delete',
    OPTIONS = 'options',
    HEAD = 'head',
    PATCH = 'patch',
    TRACE = 'trace',
  }

  export interface Request {
    body?: any
    headers?: object
    params?: object
    query?: object
  }
}

export namespace OpenAPIV3_1 {
  type Modify<T, R> = Omit<T, keyof R> & R
// ...
  export type HttpMethods = OpenAPIV3.HttpMethods

  export type PathItemObject<T extends {} = {}> = Modify<
    OpenAPIV3.PathItemObject<T>,
    {
      servers?: ServerObject[]
      parameters?: ParameterObject[]
    }
  > & {
    [method in HttpMethods]?: OperationObject<T>
  }

  export type OperationObject<T extends {} = {}> = Modify<
    OpenAPIV3.OperationObject<T>,
    {
      parameters?: ParameterObject[]
      requestBody?: RequestBodyObject
      responses?: ResponsesObject
      callbacks?: Record<string, CallbackObject>
      servers?: ServerObject[]
    }
  > &
    T

  export type ExternalDocumentationObject =
    OpenAPIV3.ExternalDocumentationObject

  export type ParameterObject = OpenAPIV3.ParameterObject

  export type HeaderObject = OpenAPIV3.HeaderObject

  export type ParameterBaseObject = OpenAPIV3.ParameterBaseObject

  export type NonArraySchemaObjectType =
    | OpenAPIV3.NonArraySchemaObjectType
    | 'null'

  export type ArraySchemaObjectType = OpenAPIV3.ArraySchemaObjectType
// ...
  export type ResponsesObject = Record<string, ResponseObject>

  export type ResponseObject = Modify<
    OpenAPIV3.ResponseObject,
    {
      headers?: { [header: string]: HeaderObject }
      content?: { [media: string]: MediaTypeObject }
      links?: { [link: string]: LinkObject }
    }
  >

  export type LinkObject = Modify<
    OpenAPIV3.LinkObject,
    {
      server?: ServerObject
    }
  >

  export type CallbackObject = Record<string, PathItemObject>
// ...
}